poprawiona konfiguracja api, dodane handlery dla PUT, DELETE

+++ server/aws_deploy.ps1
@@ -15,4 +15,6 @@ aws cloudformation deploy `
     --capabilities CAPABILITY_NAMED_IAM
 
 # cleanup
-rm $PSScriptRoot/aws-out.yaml
+if ($?) {
+    rm $PSScriptRoot/aws-out.yaml
+}
